Use knex to run seedings
